requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.1.0
astor==0.8.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.7
decorator==5.1.1
distlib==0.3.6
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
faiss-cpu==1.7.4
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google==3.0.0
google-ai-generativelanguage==0.2.0
google-api-core==1.34.0
google-auth==2.19.0
google-generativeai==0.1.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.48.2
idna==3.4
importlib-metadata==6.6.0
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.154
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
pandas-stubs==2.0.1.230501
pandasai==0.2.15
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.5.1
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==3.20.3
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.8
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rich==13.3.5
rsa==4.9
semver==3.0.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
stack-data==0.6.2
streamlit==1.18.1
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
types-pytz==2023.3.0.0
typing-inspect==0.9.0
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
validators==0.20.0
virtualenv==20.23.0
watchdog==3.0.0
wcwidth==0.2.6
yarg==0.1.9
yarl==1.9.2
zipp==3.15.0